requirements.txt
gitpython==3.1.42
python-dotenv==1.0.1
typer==0.9.0
rich==13.7.0 